Introduce dex services
