+++ src/service/dex/offers_graph.ts
@@ -1,17 +1,54 @@
 import { BigNumber } from "bignumber.js";
-import { AssetID } from "../model";
-import { Offer } from "../orm/entities";
-import logger from "../util/logger";
-import { IOrder, OrderBook } from "./orderbook";
+import { AssetID } from "../../model";
+import { Offer } from "../../orm/entities";
+import logger from "../../util/logger";
 
 interface IEdgeData {
   capacity: BigNumber;
-  orderBook: OrderBook;
+  orderBook: AssetOrders;
 }
 
 interface IEdge {
   vertex: AssetID;
   data: IEdgeData;
+}
+
+interface IAssetOrder {
+  amount: BigNumber;
+  price: BigNumber;
+}
+
+class AssetOrders {
+  constructor(private readonly orders: IAssetOrder[] = []) {}
+
+  public getOrders() {
+    // return copy of the orders
+    return this.orders.map(o => o);
+  }
+
+  public addOrder(order: IAssetOrder): void {
+    this.orders.push(order);
+  }
+
+  public sort() {
+    this.orders.sort((a, b) => a.price.comparedTo(b.price));
+  }
+
+  public buy(amountToBuy: BigNumber): BigNumber {
+    let amountToSell = new BigNumber(0);
+
+    for (const { amount, price } of this.orders) {
+      if (amountToBuy.gt(amount)) {
+        amountToSell = amountToSell.plus(amount.times(price));
+        amountToBuy = amountToBuy.minus(amount);
+      } else {
+        amountToSell = amountToSell.plus(amountToBuy.times(price));
+        break;
+      }
+    }
+
+    return amountToSell;
+  }
 }
 
 export class OffersGraph {
@@ -33,12 +70,12 @@ export class OffersGraph {
       const [assetToBuy, assetToSell] = [offer.buying, offer.selling];
 
       const edge = this.getEdgeData(assetToSell, assetToBuy);
-      const order: IOrder = { amount: offer.amount, price: offer.price };
+      const order: IAssetOrder = { amount: offer.amount, price: offer.price };
 
       if (!edge) {
         this.addEdge(assetToSell, assetToBuy, {
           capacity: offer.amount,
-          orderBook: new OrderBook([order])
+          orderBook: new AssetOrders([order])
         });
       } else {
         edge.orderBook.addOrder(order);
@@ -60,7 +97,7 @@ export class OffersGraph {
     }
 
     let capacity = new BigNumber(0);
-    const orderBook = new OrderBook();
+    const orderBook = new AssetOrders();
 
     for (const { amount, price } of offers) {
       orderBook.addOrder({ amount, price });
