+++ src/service/dex/path_finder.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from "bignumber.js";
-import { AssetID } from "../model";
-import { OffersGraph } from "./data_structure";
+import { AssetID } from "../../model";
+import { OffersGraph } from "./offers_graph";
 
 interface IPaths {
   [sourceAsset: string]: { amountNeeded: BigNumber; path: AssetID[] };
